#3: adjusted all deps and can now start but still one crash to be solve - runtime
