+++ assets/emotions.json
