+++ package.json
@@ -1,51 +1,55 @@
 {
-    "name": "papanache",
-    "version": "1.0.0",
-    "description": "Stylish Web Packer",
-    "scripts": {
-      "test": "savor test",
-      "coverage": "savor coverage"
-    },
-    "repository": {
-      "type": "git",
-      "url": "git+https://github.com/fluidtrends/papanache.git"
-    },
-    "homepage": "https://github.com/fluidtrends/papanache",
-    "dependencies": {
-        "@babel/cli": "^7.8.4",
-        "url-loader": "0.5.9",
-        "react-css-modules": "^4.7.1",
-        "@babel/preset-env": "^7.8.4",
-        "@babel/preset-react": "^7.8.3",
-        "@hot-loader/react-dom": "^16.11.0",
-        "autoprefixer": "7.1.1",
-        "copyfiles": "^1.2.0",
-        "cross-env": "^3.1.4",
-        "ora": "^1.3.0",
-        "babel-plugin-import": "^1.13.0",
-        "@babel/plugin-proposal-class-properties": "^7.8.3",
-        "@babel/plugin-proposal-object-rest-spread": "^7.8.3",
-        "@babel/plugin-transform-destructuring": "^7.8.3",
-        "@babel/plugin-transform-react-jsx": "^7.8.3",
-        "case-sensitive-paths-webpack-plugin": "2.1.1",
-        "sw-precache-webpack-plugin": "0.11.3",
-        "uglifyjs-webpack-plugin": "^1.2.4",
-        "compression-webpack-plugin": "^1.0.1",
-        "copy-webpack-plugin": "^4.0.1",
-        "static-site-generator-webpack-plugin": "^3.4.1",
-        "extract-text-webpack-plugin": "^2.1.2",
-        "html-webpack-plugin": "^3.2.0",
-        "babel-loader": "^8.0.6",
-        "sass-loader": "^6.0.6",
-        "react-hot-loader": "^4.12.19",
-        "style-loader": "^0.18.2",
-        "responsive-loader": "^1.1.0",
-        "file-loader": "0.11.2",
-        "html-loader": "^0.5.1",
-        "image-loader": "0.0.1"
-    },
-    "devDependencies": {
-      "savor": "^1.1.2"
-    }
+  "name": "papanache",
+  "version": "1.0.0",
+  "description": "Stylish Web Packer",
+  "scripts": {
+    "test": "savor test",
+    "coverage": "savor coverage"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/fluidtrends/papanache.git"
+  },
+  "homepage": "https://github.com/fluidtrends/papanache",
+  "dependencies": {
+    "@babel/cli": "^7.8.4",
+    "@babel/plugin-proposal-class-properties": "^7.8.3",
+    "@babel/plugin-proposal-object-rest-spread": "^7.8.3",
+    "@babel/plugin-transform-destructuring": "^7.8.3",
+    "@babel/plugin-transform-react-jsx": "^7.8.3",
+    "@babel/preset-env": "^7.8.4",
+    "@babel/preset-react": "^7.8.3",
+    "@hot-loader/react-dom": "^16.11.0",
+    "autoprefixer": "7.1.1",
+    "babel-loader": "^8.0.6",
+    "babel-plugin-import": "^1.13.0",
+    "case-sensitive-paths-webpack-plugin": "2.1.1",
+    "compression-webpack-plugin": "^1.0.1",
+    "copy-webpack-plugin": "^4.0.1",
+    "copyfiles": "^1.2.0",
+    "cross-env": "^3.1.4",
+    "ejs": "^3.0.1",
+    "extract-text-webpack-plugin": "^2.1.2",
+    "file-loader": "0.11.2",
+    "html-loader": "^0.5.1",
+    "html-webpack-plugin": "^3.2.0",
+    "image-loader": "0.0.1",
+    "ora": "^1.3.0",
+    "react-css-modules": "^4.7.1",
+    "react-hot-loader": "^4.12.19",
+    "responsive-loader": "^1.1.0",
+    "sass-loader": "^6.0.6",
+    "static-site-generator-webpack-plugin": "^3.4.1",
+    "style-loader": "^0.18.2",
+    "styled-jsx": "^3.2.4",
+    "sw-precache-webpack-plugin": "0.11.3",
+    "uglifyjs-webpack-plugin": "^1.2.4",
+    "url-loader": "0.5.9",
+    "webpack": "^4.42.0",
+    "webpack-dev-server": "^3.10.3"
+  },
+  "devDependencies": {
+    "react": "^16.13.0",
+    "savor": "^1.1.2"
   }
-  
+}
