+++ src/web/staticPlugin.js
