--- test/specs/main.js
@@ -1,16 +1,0 @@
-/* eslint-disable no-unused-expressions */
-
-const savor = require('savor')
-const Main = require('../../src/Main')
-
-savor.
-
-add('initial', (context, done) => {
-    const main = new Main({ test: "test1234" })
-    context.expect(main.props.test).to.equal("test1234")
-    
-    done()
-}).
-
-
-run('[Carmel Blockchain] main')
